pyproject.toml
[project]
name = "transformers-en-ko-dataset-harness"
version = "0.1.0"
description = "Codex-oriented harness for building an EN-to-KO translation dataset from multilingual documentation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "GitPython>=3.1.43",
  "python-frontmatter>=1.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.5",
]
translation = [
  "accelerate>=1.10.0",
  "datasets>=4.1.1",
  "evaluate>=0.4.6",
  "huggingface-hub>=0.34.0",
  "sacrebleu>=2.5.1",
  "sentencepiece>=0.2.0",
  "transformers>=4.57.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
